fix(api): adjust data type for balance credit limit property

+++ packages/webtrit_api/lib/src/models/common.dart
@@ -26,7 +26,7 @@ class Balance with _$Balance {
   const factory Balance({
     BalanceType? balanceType,
     double? amount,
-    int? creditLimit,
+    double? creditLimit,
     String? currency,
   }) = _Balance;
